pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "profile-backend"
version = "0.1.0"
description = "Local profile PDF/DOCX pipeline"
requires-python = ">=3.11"
dependencies = [
  "pdfplumber>=0.11.0",
  "python-docx>=1.1.0",
  "spacy>=3.7.0",
  "openpyxl>=3.1.0",
  "Flask>=3.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["profile_backend*"]
